refactor(app): move the sub-agent spawner into internal/app (PR1b-2)
The Spawner (agentSpawner) builds isolated child agents for launch_agent /
send_message. It depended only on agent + tools, so it relocates cleanly to
internal/app next to the sender and gate — the shared building blocks the
bootstrap will assemble.

- internal/app.NewSpawner(parent, executor, toolsFn) returns the tools.Spawner
  (renamed from cmd/octo's unexported agentSpawner → exported Spawner). The
  child registry, tool filtering, and delta token accounting move with it,
  verbatim.
- cmd/octo (chat, conduct) call app.NewSpawner; the spawner tests (unit +
  the send_message-resume integration test) move to internal/app, where they
  keep access to the unexported registry internals they assert on.

Behavior-preserving — full suite + race green, vet + gofmt clean. Next:
app.Bootstrap wires sender + agent knobs + executor + this spawner + sub-agent
manager + task store + tool-search + gate, and the CLI migrates onto it.
